+++ aframe/src/location-based/gps-camera.js
@@ -39,7 +39,8 @@ AFRAME.registerComponent('gps-camera', {
         this.loader = document.createElement('DIV');
         this.loader.classList.add('arjs-loader');
         document.body.appendChild(this.loader);
-        window.addEventListener('gps-entity-place-added', function() {
+
+        window.addEventListener('gps-entity-place-added', function () {
             // if places are added after camera initialization is finished
             if (this.originCoords) {
                 window.dispatchEvent(new CustomEvent('gps-camera-origin-coord-set'));
@@ -60,13 +61,13 @@ AFRAME.registerComponent('gps-camera', {
         if (!!navigator.userAgent.match(/Version\/[\d.]+.*Safari/)) {
             // iOS 13+
             if (typeof DeviceOrientationEvent.requestPermission === 'function') {
-                var handler = function() {
+                var handler = function () {
                     console.log('Requesting device orientation permissions...')
                     DeviceOrientationEvent.requestPermission();
                     document.removeEventListener('touchend', handler);
                 };
 
-                document.addEventListener('touchend', function() { handler() }, false);
+                document.addEventListener('touchend', function () { handler() }, false);
 
                 alert('After camera permission prompt, please tap the screen to activate geolocation.');
             } else {
@@ -207,7 +208,7 @@ AFRAME.registerComponent('gps-camera', {
             this._setPosition();
         }
     },
-    _setPosition: function() {
+    _setPosition: function () {
         var position = this.el.getAttribute('position');
 
         // compute position.x
@@ -215,6 +216,7 @@ AFRAME.registerComponent('gps-camera', {
             longitude: this.currentCoords.longitude,
             latitude: this.originCoords.latitude,
         };
+
         position.x = this.computeDistanceMeters(this.originCoords, dstCoords);
         position.x *= this.currentCoords.longitude > this.originCoords.longitude ? 1 : -1;
 
@@ -223,11 +225,15 @@ AFRAME.registerComponent('gps-camera', {
             longitude: this.originCoords.longitude,
             latitude: this.currentCoords.latitude,
         }
+
         position.z = this.computeDistanceMeters(this.originCoords, dstCoords);
         position.z *= this.currentCoords.latitude > this.originCoords.latitude ? -1 : 1;
 
         // update position
         this.el.setAttribute('position', position);
+
+
+        window.dispatchEvent(new CustomEvent('gps-camera-update-position', { detail: { position: this.currentCoords, origin: this.originCoords }}));
     },
     /**
      * Returns distance in meters between source and destination inputs.